requirements.txt
# Backend
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
python-multipart>=0.0.12
aiosqlite>=0.20.0
authlib>=1.3.0
itsdangerous>=2.2.0

# Document processing
pymupdf>=1.26.0
pdfplumber>=0.11.4
pillow>=10.4.0
python-pptx>=0.6.23
python-docx>=1.1.2
rapidfuzz>=3.9.6
spacy>=3.8.0
pytesseract>=0.3.10

# NLP model (download separately: python -m spacy download en_core_web_md)
# en_core_web_md
